Reset to clean 1v1 and add Destroyer bot curriculum
