+++ bot.cfg
@@ -4,13 +4,13 @@ num_participants = 2
 [Participant 0]
 team = 0
 type = rlbot
-bot_name = Necto
+bot_name = Destroyer
 python_file = bot.py
 loadout_config = appearance.cfg
 
 [Participant 1]
 team = 1
 type = rlbot
-bot_name = Necto
+bot_name = Destroyer
 python_file = bot.py
 loadout_config = appearance.cfg
