+++ requirements.txt
@@ -1,13 +1,10 @@
 # Include everything the framework requires
-# You will automatically get updates for all versions starting with "1.".
-rlbot==1.*
+rlbot>=1.74.0
 --find-links https://download.pytorch.org/whl/torch_stable.html
-torch==2.0.1+cpu
+torch>=2.1
 rlgym-compat==1.0.2
-numpy
+numpy>=1.24
+gymnasium>=0.29
 
 # This will cause pip to auto-upgrade and stop scaring people with warning messages
 pip
-torch>=2.1
-numpy>=1.24
-gymnasium>=0.29
