+++ trainer_online_bc.py
@@ -1,28 +1,63 @@
-import time, torch, numpy as np
-from torch.utils.data import DataLoader, TensorDataset
+import time, numpy as np
+try:
+    import torch
+    TORCH_OK = True
+except Exception:
+    TORCH_OK = False
 
-def train_step(snapshot, model, opt, batch=256, epochs=2):
-    obs, act, _, _ = snapshot
-    if len(obs) < batch: return None
-    ds = TensorDataset(torch.from_numpy(obs), torch.from_numpy(act))
-    dl = DataLoader(ds, batch_size=batch, shuffle=True, drop_last=True)
-    tot=0.0; n=0
-    for _ in range(epochs):
-        for o,a in dl:
-            pred = model(o)
-            loss = ((pred - a)**2).mean()
-            opt.zero_grad(); loss.backward()
-            torch.nn.utils.clip_grad_norm_(model.parameters(), 1.0)
-            opt.step()
-            tot += loss.item(); n += 1
-    return tot/max(n,1)
+class OnlineBC:
+    def __init__(self, buffer, policy, out_path="necto-model.pt", step_every=3.0, batch=256, lr=2e-4):
+        self.buffer = buffer
+        self.policy = policy
+        self.out_path = out_path
+        self.step_every = step_every
+        self.batch = batch
+        self.lr = lr
+        self.stop_flag = False
 
-def online_loop(buffer, model, optimizer, out_path="necto-model.pt", step_every=3.0):
-    while True:
-        time.sleep(step_every)
-        snap = buffer.snapshot()
-        loss = train_step(snap, model, optimizer)
-        if loss is None: continue
-        scripted = torch.jit.trace(model, torch.randn(1, snap[0].shape[1]))
-        scripted.save(out_path)
-        print(f"[online_bc] loss={loss:.4f} saved {out_path}")
+    def start(self):
+        import threading
+        threading.Thread(target=self.run, daemon=True).start()
+
+    def run(self):
+        if not TORCH_OK:
+            return
+        self.policy.maybe_reload()
+        model = self.policy.model
+        if model is None:
+            return
+        try:
+            opt = torch.optim.Adam(model.parameters(), lr=self.lr)
+        except Exception:
+            return
+
+        last_save = time.time()
+        while not self.stop_flag:
+            time.sleep(self.step_every)
+            obs, act, _, _ = self.buffer.snapshot()
+            if len(obs) < self.batch:
+                continue
+            try:
+                x = torch.from_numpy(obs).float()
+                y = torch.from_numpy(act).float()
+                idx = np.random.choice(len(x), size=min(self.batch, len(x)), replace=False)
+                xb, yb = x[idx], y[idx]
+                pred = model(xb)
+                if pred.shape[-1] != 8:
+                    continue
+                loss = ((pred - yb) ** 2).mean()
+                opt.zero_grad(); loss.backward()
+                torch.nn.utils.clip_grad_norm_(model.parameters(), 1.0)
+                opt.step()
+            except Exception:
+                continue
+
+            if time.time() - last_save > self.step_every:
+                try:
+                    torch.jit.save(model, self.out_path)
+                    last_save = time.time()
+                except Exception:
+                    pass
+
+    def stop(self):
+        self.stop_flag = True
